Correccion de entidades

+++ examen/src/main/java/org/una/examen/entities/Unidad.java
@@ -49,7 +49,7 @@ public class Unidad {
     private Integer poblacion;
     
     @Column
-    private Float area;
+    private Double area;
     
     @ManyToOne 
     @JoinColumn(name="distrito")
